requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0
pyyaml>=6.0
boto3>=1.28.0
s3fs==2025.3.0
fsspec==2025.3.0
aiobotocore>=2.5.4
pytest>=7.4.0
pytest-cov>=4.1.0